Move core database_configuration patch into module
Takes the core code that is in an anonymous module in the
database_configuration.rb file and move it into it's own file so it can
be reused in otherplaces as well.
